# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
      "Body_upload_files_api_v1_sources_uploads_post": {
        "properties": {
          "files": {
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          }
        },
        "required": [
          "files"
        ],
        "title": "Body_upload_files_api_v1_sources_uploads_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PatchSourceRequest": {
        "properties": {
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "active",
                  "paused"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "PatchSourceRequest",
        "type": "object"
      },
...
      "RegisterSourceRequest": {
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "kind": {
            "const": "folder",
            "default": "folder",
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "uri": {
            "minLength": 1,
            "title": "Uri",
            "type": "string"
          }
        },
        "required": [
          "name",
          "uri"
        ],
        "title": "RegisterSourceRequest",
        "type": "object"
      },
      "RegisterSourceResponse": {
        "properties": {
          "batch_id": {
            "title": "Batch Id",
            "type": "string"
          },
          "enqueued": {
            "title": "Enqueued",
            "type": "integer"
          },
          "skipped_unchanged": {
            "title": "Skipped Unchanged",
            "type": "integer"
          },
          "source": {
            "$ref": "#/components/schemas/SourceView"
          }
        },
        "required": [
          "source",
          "batch_id",
          "enqueued",
          "skipped_unchanged"
        ],
        "title": "RegisterSourceResponse",
        "type": "object"
      },
      "ReindexResponse": {
        "properties": {
          "batch_id": {
            "title": "Batch Id",
            "type": "string"
          },
          "deduplicated": {
            "title": "Deduplicated",
            "type": "integer"
          },
          "enqueued": {
            "title": "Enqueued",
            "type": "integer"
          },
          "skipped_unchanged": {
            "title": "Skipped Unchanged",
            "type": "integer"
          }
        },
        "required": [
          "batch_id",
          "enqueued",
          "skipped_unchanged",
          "deduplicated"
        ],
        "title": "ReindexResponse",
        "type": "object"
      },
      "RejectedFileView": {
        "properties": {
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "filename",
          "reason"
        ],
        "title": "RejectedFileView",
        "type": "object"
      },
...
      "SourceView": {
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "last_indexed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Indexed At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "uri": {
            "title": "Uri",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "name",
          "uri",
          "status",
          "config",
          "last_indexed_at",
          "created_at"
        ],
        "title": "SourceView",
        "type": "object"
      },
      "UploadResponse": {
        "properties": {
          "batch_id": {
            "title": "Batch Id",
            "type": "string"
          },
          "enqueued": {
            "title": "Enqueued",
            "type": "integer"
          },
          "rejected": {
            "items": {
              "$ref": "#/components/schemas/RejectedFileView"
            },
            "title": "Rejected",
            "type": "array"
          },
          "skipped_unchanged": {
            "title": "Skipped Unchanged",
            "type": "integer"
          },
          "source": {
            "$ref": "#/components/schemas/SourceView"
          },
          "stored": {
            "items": {
              "type": "string"
            },
            "title": "Stored",
            "type": "array"
          }
        },
        "required": [
          "source",
          "batch_id",
          "stored",
          "rejected",
          "enqueued",
          "skipped_unchanged"
        ],
        "title": "UploadResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Atlas API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/sources": {
      "get": {
        "operationId": "list_sources_api_v1_sources_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SourceView"
                  },
                  "title": "Response List Sources Api V1 Sources Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Sources",
        "tags": [
          "sources"
        ]
      },
      "post": {
        "operationId": "register_source_api_v1_sources_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterSourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterSourceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Source",
        "tags": [
          "sources"
        ]
      }
    },
    "/api/v1/sources/uploads": {
      "post": {
        "description": "Store dropped files in the managed source and index them.",
        "operationId": "upload_files_api_v1_sources_uploads_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_files_api_v1_sources_uploads_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Files",
        "tags": [
          "sources"
        ]
      }
    },
    "/api/v1/sources/{source_id}": {
      "delete": {
        "operationId": "delete_source_api_v1_sources__source_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Source",
        "tags": [
          "sources"
        ]
      },
      "patch": {
        "operationId": "patch_source_api_v1_sources__source_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchSourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Source",
        "tags": [
          "sources"
        ]
      }
    },
    "/api/v1/sources/{source_id}/reindex": {
      "post": {
        "operationId": "reindex_source_api_v1_sources__source_id__reindex_post",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReindexResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reindex Source",
        "tags": [
          "sources"
        ]
      }
    },
...
  }
}